Create test base class
